Introduce tarerofs I/O API
Previously, erofs-utils was integrated into overlaybd by
using the `mkfs.erofs` executable file. It's not optimial
since raw data needs to be dumped first and output data
needs to be write into overlaybd then.

This commit introduces the I/O API in tarerofs, that the data
stream of the:
- output image
- input tar file
can be controlled by the I/O manager.

This also adds block read/write cache support for IFile to
improve the efficiency of making file system images.

Signed-off-by: Hongzhen Luo <hongzhen@linux.alibaba.com>